backend/pyproject.toml
[project]
name = "podcast-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.114.0",
    "nltk>=3.9.1",
    "python-dotenv>=1.1.0",
    "torch>=2.6.0",
    "transformers>=4.51.2",
    "youtube-transcript-api>=1.0.3",
]
